Added Chef Cookbook Install Code

+++ Jenkinsfile
@@ -8,20 +8,22 @@ pipeline {
     }
     stage("Installing Chef Workstation") {
       steps {
-        sh 'sudo apt-get install -y wget tree unzip'
-        sh 'wget https://packages.chef.io/files/stable/chef-workstation/0.15.6/ubuntu/18.04/chef-workstation_0.15.6-1_amd64.deb'
-        sh 'sudo dpkg -i chef-workstation_0.15.6-1_amd64.deb'
-        sh 'sudo chef env --chef-license accept'
+        script {
+          def exists = fileExists '/usr/bin/chef-client'
+          if (exists) {
+            echo "Skipping Chef Workstation Install'
+          } else {
+            echo "Installing Chef Workstation"
+            sh 'sudo apt-get install -y wget tree unzip'
+            sh 'wget https://packages.chef.io/files/stable/chef-workstation/0.15.6/ubuntu/18.04/chef-workstation_0.15.6-1_amd64.deb'
+            sh 'sudo dpkg -i chef-workstation_0.15.6-1_amd64.deb'
+            sh 'sudo chef env --chef-license accept'
+         }
       }
     }
-    stage('Second Stage') {
+    stage('Download Apache Cookbook') {
       steps {
-        echo 'Second stage'
-      }
-    }
-    stage('Third Stage') {
-      steps {
-        echo 'Third stage'
+        git credentialsId: 'git-repo-id', url: 'git@github.com:jeffmbelt/apache.git'
       }
     }
   }
